service/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.5.3
numpy==1.26.4